Enable zsh by default
To fix this:

```
Failed assertions:

- users.users.nikita.shell is set to zsh, but
programs.zsh.enable is not true. This will cause the zsh
shell to lack the basic nix directories in its PATH and might make
logging in as that user impossible. You can fix it with:
programs.zsh.enable = true;

If you know what you're doing and you are fine with the behavior,
set users.users.nikita.ignoreShellProgramCheck = true;
instead.
```